requirements_min.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1
eventlet==0.33.3
gunicorn==21.2.0
pyautogui==0.9.54
Pillow==10.1.0
pyperclip==1.8.2